Reorganize redirection `Policy` combinators
Commit 04cc691e034342be605242b33f0ff976e8316d2a added a new type
parameter to `Policy` trait to represent the error type and, as a side
effect, complicated type inference when calling the trait methods. To
mitigate the issue, this commit moves the `Policy` combinators from
trait methods to freestanding functions.

This also adds `select` combinator function, which is useful in
combination with `Action::error`.