Change compression body error type to `BoxError`
The compression related middleware would previously use `BodyOrIoError`
as the error type. That only implemented `std::error::Error` if the
inner error did as well. Since `Box<dyn std::error::Error + Send +
Sync>` does not implement `std::error::Error` `Compression` and
`Decompression` wouldn't be usable with hyper if the body they wrapped
had that error type.

This changes the middleware to use `BoxError` as the error type which
resolves the issue. Its also consistent with other middleware.

Note: This is a breaking change!