Allow users to control level of Tracing spans (#124)
* Manually implement `Default` for `DefaultMakeSpan`

This doesn't make sense as-is, but I'm about to add an additional
non-`Default` field to `DefaultMakeSpan`. It will no longer be possible
to derive `Default` once that happens.

* Make log level used for trace spans configurable

This is a minimally invasive fix for #123. It's certainly not ideal,
API-wise, but at least it allows users to add request information to
their logs without having to set the log level to DEBUG.

Close #123

* Simplify doc links

Those no longer need to be speicified explicitly in recent versions of
rustdoc.

* Point out potential pitfall in documentation

* Add changelog entry

Co-authored-by: David Pedersen <david.pdrsn@gmail.com>