feat(fs): make AsyncReadBody::with_capacity() pub
The `AsyncReadBody` type is a generally useful type. It is already public, but it has no public constructor. This commit makes the main constructor public so that the type can be used by other projects. If the type is ever upstreamed to `http-body`, it can be reexported here to maintain compatibility.