Patch tower-http:0.3 to come from the current workspace
This avoids pulling in a duplicate remote copy of tower-http in the
examples.

This patch will need to be updated any time we bump our minor version.
We could patch it unconditionally instead but that risks breaking our
examples if they run into any of our backwards-incompatible changes. If
we forget to update this, `cargo deny` should catch it for us.