New `gateway` module/feature
This module provides a gateway/reverse proxy service. For any input
request, it rewrites the URL to be relative to a configured base URL,
processes the `Connection` header, and removes any other standard
hop-by-hop headers[^1]. It also adds `Forwarded`, `Via`, and optionally
`X-Forwarded-*` headers if configured with a getter for information
about the connection.

The intended use here is to wrap something like `hyper::Client` as the
underlying transport, though this service supports any transport that
accepts an `http::Request` regardless of the underlying protocol it
uses.

This service provides `Forwarded` and `Via` headers by default (with
appropriate configuration) as those are the standardized headers.
Support for `X-Forwarded-*` headers are opt-in as these are non-standard
headers despite being very commonly-used.

[^1]: It leaves `TE`, `Transfer-Encoding`, and `Trailer` intact despite
these being hop-by-hop headers, with the rationale that it does not
modify the request or response body. If the server modifies the request
or response body in another layer it should modify these headers
appropriately. This is explained in the module docs.