Make type inference work for ServiceExt
Unfortunately, making type inference work for this requires removing the
generic parameter from ServiceExt, which in turn requires removing the
tower::Service supertrait. This made the Sealed trait useless, and while
I was at it I also removed the Sized supertrait in favor of Sized bounds
on all current trait methods (this allows backwards-compatibly adding
methods that don't require Self to be Sized to be added).

This means that now you can call all of these methods on _anything_, the
resulting type just isn't going to be very useful if the type you're
calling the method on isn't a tower::Service<http::Request<B>> (extra
bounds may apply depending on the exact method).